Split Google Health providers into file-per-data-type
Addresses review: replace the single google-health.ts with one file per data
type (grouping sub-type families like sleep, heart-rate min/max/avg, activity
minutes and HR-zone minutes, matching the Fitbit/Garmin layout), backed by a
shared google-health-common.ts query helper. Also removes the back-end
type-name docstring per review. Barrel exports and all consumers are unchanged.
