+++ src/helpers/daily-data-providers/google-health-spo2.ts
@@ -1,0 +1,6 @@
+import { DailyDataQueryResult } from "../query-daily-data";
+import { googleHealthDailyValue } from "./google-health-common";
+
+export default function (startDate: Date, endDate: Date): Promise<DailyDataQueryResult> {
+    return googleHealthDailyValue("dailyOxygenSaturation-list-avg", startDate, endDate);
+}
