+++ src/helpers/daily-data-providers/index.ts
@@ -93,6 +93,31 @@ export { default as ouraStepsDataProvider } from "./oura-daily-steps"
 export { default as ouraSleepMinutesDataProvider } from "./oura-total-sleep"
 export { default as ouraRestingHeartRateDataProvider } from "./oura-resting-heart-rate"
 export { default as ouraActiveCaloriesBurnedDataProvider } from "./oura-active-calories-burned"
-export * from "./google-health"
+export { default as googleHealthStepsDataProvider } from "./google-health-steps"
+export { default as googleHealthRestingHeartRateDataProvider } from "./google-health-resting-heart-rate"
+export { default as googleHealthCaloriesBurnedDataProvider } from "./google-health-calories-burned"
+export { default as googleHealthActiveCaloriesBurnedDataProvider } from "./google-health-active-calories-burned"
+export { default as googleHealthFloorsDataProvider } from "./google-health-floors"
+export { default as googleHealthDistanceDataProvider } from "./google-health-distance"
+export { default as googleHealthWearMinutesDataProvider } from "./google-health-wear-minutes"
+export { default as googleHealthBreathingRateDataProvider } from "./google-health-breathing-rate"
+export { default as googleHealthHrvDataProvider } from "./google-health-hrv"
+export { default as googleHealthSpO2DataProvider } from "./google-health-spo2"
+export { maxHeartRate as googleHealthMaxHeartRateDataProvider } from "./google-health-heart-rate"
+export { minHeartRate as googleHealthMinHeartRateDataProvider } from "./google-health-heart-rate"
+export { averageHeartRate as googleHealthAverageHeartRateDataProvider } from "./google-health-heart-rate"
+export { sedentaryMinutes as googleHealthSedentaryMinutesDataProvider } from "./google-health-activity-minutes"
+export { totalActiveMinutes as googleHealthTotalActiveMinutesDataProvider } from "./google-health-activity-minutes"
+export { lightlyActiveMinutes as googleHealthLightlyActiveMinutesDataProvider } from "./google-health-activity-minutes"
+export { fairlyActiveMinutes as googleHealthFairlyActiveMinutesDataProvider } from "./google-health-activity-minutes"
+export { veryActiveMinutes as googleHealthVeryActiveMinutesDataProvider } from "./google-health-activity-minutes"
+export { totalElevatedHeartRateMinutes as googleHealthElevatedHeartRateMinutesDataProvider } from "./google-health-elevated-heart-rate"
+export { fatBurnMinutes as googleHealthFatBurnMinutesDataProvider } from "./google-health-elevated-heart-rate"
+export { cardioMinutes as googleHealthCardioMinutesDataProvider } from "./google-health-elevated-heart-rate"
+export { peakMinutes as googleHealthPeakMinutesDataProvider } from "./google-health-elevated-heart-rate"
+export { totalSleepMinutes as googleHealthTotalSleepMinutesDataProvider } from "./google-health-sleep"
+export { lightSleepMinutes as googleHealthLightSleepMinutesDataProvider } from "./google-health-sleep"
+export { remSleepMinutes as googleHealthRemSleepMinutesDataProvider } from "./google-health-sleep"
+export { deepSleepMinutes as googleHealthDeepSleepMinutesDataProvider } from "./google-health-sleep"
 export * from "./google-health-preferred-over-fitbit"
 export * from "./shared"
